Change VC project settings to statically include SFML
Minimizes the mess from when I originally packed this project. Neither
SFML nor this game are big enough to make this a big deal on size.

A remaining mess is the settings for include/lib directories - edited to
relative paths from SolutionDir, but the dir hierarchy is still awful
(SFML folder inside SFML folder and must be 2 levels above this project
folder). Apologies to any other person (or. PC, even) who might want to
work with this source code.